requirements.txt
python-telegram-bot==20.7
supabase==2.4.1
python-dotenv==1.0.0
requests==2.31.0
